Persist items into a SQLite DB instead of to the filesystem.
Follows the same methodology as before. Whenever there's a change to the list,
the entire list is written. Every item is deleted from the DB, and then each
item is rewritten. This makes it easy to persist the position of each item in
the list. The tradeoff is that it is inefficient.